British Dominion release events are retuned to be less spammy
This is done through the addition of a blank country modifier that enforces a 3 year cooldown. This cooldown agrees approximately with the time between historical British Colonial Conferences. If a dominion is released, the cooldown is not incurred. If release is refused, the cooldown is incurred. Some of the events, especially Canadian, are retuned to take into account historical factors leading to the concept of British Dominions. Some militancy and consciousness effects of release refusal are toned down to be more realistic given the relatively orderly establishment of British dominions historically.

Added the French dominion events for Canada, Louisiana, and French New Zealand as well.