Event 99101 is changed to occur when the player is Afghanistan
Event 99101 was problematic because there was no reason for the player to expect it to be ai-only. I have changed this to no longer be the case. Additionally, I have added a new events file to provide notifications to tags that experience ENG flavor without having previously been notified by it. I expect to use this file for other missing notifications, for the Trucial States and other tags that are impacted by older flavor code, especially.