Several minor fixes to French flavor
Fixing the hidden infamy cost of event 99949, The Death of Henri Riviere, led to a search through French flavor to see if other hidden infamy costs existed. None were found, but many minor issues were found instead.

Rename Louisiana remained allowed after being clicked, because it seemed to be incomplete. The rename Washington flags should have been rename Montana instead, suggesting that the other states were probably planned to be renamed as well, but never were.

Event 95650, meant to give France a casus belli against a target that held part of historical French Indochina, did not award a casus belli when only a province with a Cambodian core was left in the possession of Siam. This seems to have been because the casus belli, demand_concession_BC_casus_belli, was not valid for this. However, demand_concession_casus_belli works. This event has other major issues in balance, since it can be used to annex most of Thailand, but balancing problems are outside the scope of this set of minor fixes.

Louisiana was misspelled in a few places in localisation.