Supplies new options to reform certain formable tags after annexation, and if the tag fulfills prior requirements for the formable. Provides new options for Scandinavia to revert back to a pre-Scandinavia tag if it lacks at least two of the three mainland European capitals involved. Provided a new option for an Indonesian tag to form Indonesia after it has been organized. Other formable tags affected: Transcaucasia, Turkestan. Formables that are rarely or never formed by the ai are not addressed.
Added a requirement that the tag itself cannot see the decision that erases its formable flag

Added a fix for Tukulor Empire formation. I had thought during testing that it could be reformed after annexation, but obviously that is not true, given the check for formation in the Potential section of the decision. I moved this to the Allow section so that the player can see the possibility of reforming the Tukulor Empire even after it has been formed.